requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
kafka-python==2.0.2
confluent-kafka==2.2.0
boto3==1.28.38
moto==4.2.0
pandas==2.0.3
pyarrow==12.0.1
fastparquet==2023.7.0
numpy==1.24.4
pydantic==2.3.0
python-dotenv==1.0.0
aiobotocore==2.5.4 